Update Helm Chart version BEFORE `helm-docs` and `helm template` commands.
This corrects inconsistent behavior of the releasing script.

The order that was previously dictated by the `RELEASE_CHECKLIST.md` was
very intentional since the Helm chart provides the set of instructions (and
versions) that `helm-docs` uses to update its README and Chart examples.

The contents of the Chart also inform the output produced by `helm template`
which is inserted into the `kubernetes.mdx` file by the `update_docs`
function.