Do a pre-flight check which asserts availabilty of necessary tools
By checking for presence of required developer tooling ahead of making any
changes, we improve the developer experience and prevent a half-finished
release.

Previously, the steps in the releasing process would proceed as if you had
already ensured all of the required development tooling was installed on
your local machine - for example, having `helm` and `helm-docs`.

Therefore, during the release process, the lack of these commands present in
the user's PATH would cause a failre, even though previous steps which had
been completed were not idempotent.  This resulted in a half-finished
release which would need to be manually aborted which varied in difficulty
depending on how far the release proces was able to get.

For exmaple, the lack of `cargo-about` or `cargo-deny` meant that the entire
run would fail after it had already changed the Cargo.toml files and swapped
the CHANGELOG entries around.

The instructions produced by this check also give links to README files on
how to install the commands.