Remove pre-determined version heading from `NEXT_CHANGELOG.md`
I don't think it makes sense to have the version pre-populated in the
`NEXT_CHANGELOG.md` file prior to the release engineer making a
determination about what the release will be.

Practically speaking, it's necessary to evaluate whether a release has bug
fixes or features prior to picking a version and the picking of a version is
done by passing the appropriate paramter (patch, minor, etc.) to the release
xtask script.