Repair logic which migrates `NEXT_CHANGELOG.md` entries to `CHANGELOG.md`
The previous logic seemed to leave the `NEXT_CHANGELOG.md` file in a
corrupted state.  I've updated the regular expressions, and further changed
the mechanism to not attempt to put an empty template in place within
`NEXT_CHANGELOG.md` at the time that the migration occurs.

Practically speaking, the template is already there, it's just placed in a way
that it's an HTML comment and it's merely on the developer who introduces the
first entry to the `NEXT_CHANGELOG.md` file to pick the appropriate heading
for the entry they are writing.

Requiring the developer be more intentional about picking the appropriate
heading (e.g., "Feature", "Improvement", "Experiment", etc.) should
hopefully be a fine ask and will also avoid the automated introduction of
blank headings which need to be removed later.  It should also reduce the
almost instantaneous introduction of merge conflicts to the
`NEXT_CHANGELOG.md` file.