Support ANY version string rather than just digits-dot-digits-dot-digits.
This adjusts the regular expressions we were using previously to
treat the version as an opaque value, rather than `\d+.\d+.\d+`, and in some
cases, uses a more defensive technique for matching them against
line-endings, etc.

We should aim to keep our release process flexible enough to work with
pre-release versions (e.g., 1.2.3-alpha.0).  The process has supported this
previously, and I think that keeping that functionality is something we want
to reserve.  (I believe the only constraint is the regular expressions in
place for these version replacements that this commit removes.)