Remove quotes around invocation of `helm template`'s `--set` flags
The quotes do not seem to be getting eaten up by `helm template`'s command
line parser but instead getting introduced directly to the values.  This was
observed both with the unnecessary introduction of quotes in the YAML
strings, but (worse) when seeing that the Base64'd values for the secrets
are merely Base64'd values of `"REDACTED"` rather than `REDACTED`, which
seems fundamentally wrong.

Happy to undo this commit if I'm missing something.