add a small delay before accepting again
right now the router will go in a busy loop trying to accept new
connections when it has reached the maximum number of file descriptors.
Waiting for a bit allows the rest of the router to work on something
else and possibly finish existing connections.

Unfortunately, if we are at the limit, the router will keep alternating
between accepting connections and reaching the max open file limit, so
we need another design, possibly a kind of hysteresis