Orbiter integration runs on expanded validated config rather than raw (#2398)
Fixes #2397

When we reach orbiter integration we should have a valid configuration.
This change stashes the validated and expanded config rather than the
raw config so that we don't have to worry about doing this stuff again
within the orbiter code.

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- ~~[ ] Documentation[^2] completed~~
- ~~[ ] Performance impact assessed and acceptable~~
- Tests added and passing[^3]
    - [x] Unit Tests
    - ~~[ ] Integration Tests~~
    - [x] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

Co-authored-by: bryn <bryn@apollographql.com>