infra: Stop (mis)using the GitHub Milestones feature.
While I once started using GitHub Milestones to try and actually convey what
PR landed in which released version of the Router, GitHub now does that
automatically with the advent of the "GitHub Releases" feature which
automatically puts a green callout at the bottom of a merged PR noting which
version of Apollo Router it first was published in.

As of today, the Milestones are actually being used _after_ we do a release,
which is less helpful and duplicative work.  It also has a lot of overhead
that required a lot of code to maintain.

This removes it, as discussed directly with the core team.