Fix default value parsing of integer fields of input types (#2396)
`IntValue::source_string()` can include whitespace, which causes
`i64::from_str` to fail.

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

- Compat: I’m not sure what the macro-consequences are of the router
incorrectly thinking there is no default value, so I don’t know if
anyone could (accidentally) rely on this bug
- Docs: Is such a bugfix changelog-worthy?
- Perf: I expect (but have not measured) impact, if any, to be
indistinguishable from noise

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`