Adding APQ support on client side (subgraph calls) (#2284)
As discussed in [this
issue](https://github.com/apollographql/router/issues/567), this PR
proposes the addition of APQ in the outgoing requests in SubgraphService
(APQ support is already present for incoming requests). I would
appreciate feedback on the proposed changes.

**Note:** The below working is changed as mentioned in later comments

**How it works:** The APQ implementation has two major steps to take
care of - 1. Calculating hashes, 2. Adding them to the graphql request
and retrying the request if we receive the
[PERSISTED_QUERY_NOT_FOUND](https://www.apollographql.com/docs/apollo-server/performance/apq/#:~:text=PERSISTED_QUERY_NOT_FOUND)
error.
1. The calculation of hashes happens after query plan is received in
bridge_query_planner.rs. For this, we are traversing the query plan tree
and for every fetch node's _operation_ field, we are calculating the
hash for it and storing it as _operation_hash_ field in FetchNode. This
also enables the hash value to be cached in cache_query_planner.rs along
with the query plan.
2. This happens in subgraph_service.rs - We first remove the query from
the SubgraphRequest and try the request - Then we check the extensions
map in response.errors for the key "code" and check if that is
"PERSISTED_QUERY_NOT_FOUND". If so, we retry the SubgraphRequest with
the query string included and rest of the flow remains the same.

**Problems that need help in this** - 

1. While traversing the query plan to calculate hashes as mentioned in
step 1. above, I have to create new nodes for changing the
_operation_hash_ since the FetchNode is an immutable object. Ideally we
should have been able to update just the _operation_hash_ field rather
than recreating the whole object.
2. To check if the PERSISTED_QUERY_NOT_FOUND error is received, I am not
sure whether checking the response.errors.extensions map for the "code"
key is a best / reliable way of checking whether the server did not
recognise the persisted query.