Make Orbiter code gracefully handle invalid validated_yaml (#2413)
- Fix #2412

Make the orbiter code gracefully degrade if there are issues rather than
panic.
Cleaned up configuration creation code.
Put validated schema in an option rather than an Arc.

Note that there are still expects in the code, but these should remain
as they would point to actual bugs.


**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

This doesn't affect perf or need new docs.
An integration test could have been written, but I feel this is better
covered by the unit tests.

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

Co-authored-by: bryn <bryn@apollographql.com>