infra: Change `NEXT_CHANGELOG.md` template to exclude next version number (#2408)
Practically speaking, we can't actually guess what this version number is
going to be until all the work has been collected together on `dev` and
we consider the totality of the changeset and their impact on the version
number.

Because the previous regular expression in the release script was
intentionally fragile, this also makes it less fragile since changing the
template would have previously broken it. This won't get us terribly far,
but it will at least be a bit more durable. Note, because regex, it's still
going to be fragile, but this should give us _some_ runway. :)

Follows https://github.com/apollographql/router/pull/2323