Make Configuration::default() go through centralised parse validation and expansion logic. (#2432)
- Fix #2429

There are lots of steps in producing final validated configuration, this
is handled when configuration is hydrated from a string. Previously
Configuration::default() was using the builder rather than the parse
logic.

In addition the defaulting mechanism now works directly against config
with the env variables, so prefixing no longer interferes.

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [ ] ~~Documentation[^2] completed~~
- [ ] ~~Performance impact assessed and acceptable~~
- Tests added and passing[^3]
    - [x] Unit Tests
    - [ ] ~~Integration Tests~~
    - [x] Manual Tests

**Exceptions**

No docs needed as this is not a change or new  in functionality.

*Note any exceptions here*

**Notes**

In future we should consider removing the `Configuration::builder()`. We
can leave the fake one. Goal would be that we have a single way of
creating Configuration rather than the multiple that we have right now.

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

Co-authored-by: bryn <bryn@apollographql.com>