feat: add optional Access-Control-Max-Age header to CORS plugin (#2331)
Hi there

As discussed in this
[issue](https://github.com/apollographql/router/issues/2212), the CORS
plugin in Apollo Router currently lacks setting the
Access-Control-Max-Age header. This PR adds a new option called max_age
that is used like so:
```
cors:
  max_age: 86400
```

This does not solve the Cache-Control header mentioned in issue as it's
a bit separate from CORS.

New test(s) and documentation changes haven't been done yet as I wanted
to confirm this is fine before proceeding further. If so, happy to do
them. Thanks!

(FYI I don't have the ability to assign a reviewer or anything, if we're
able to do that)

Signed-off-by: Benjamin Coenen <5719034+bnjjj@users.noreply.github.com>
Co-authored-by: Simon Sapin <simon@apollographql.com>
Co-authored-by: Benjamin Coenen <5719034+bnjjj@users.noreply.github.com>