Add (unused) query parsing based on apollo-compiler (#2411)
Continuation of https://github.com/apollographql/router/pull/2401

Similar to there, the new `Query::parse_with_hir` constructor is not
used yet during normal Router operation.
The two implementations return identical results
for all of a corpus of ~1.2 million production queries.

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

- Compat: expect identical behavior (which is supported by testing)
- Docs: no user-visible change
- Perf: expect increase of memory use during query parsing, but unknown
how significant. And parsing is short, and its results cached.
- Manual testing with
https://github.com/apollographql/harness-vs-apollo-rs/pull/1

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`