chore(deps): update to `router-bridge` containing Federation v2.3.0 (#2462)
At surface level, the motivation for this PR is to update to a version
of `router-bridge` which includes Federation 2.3.0 — available as of
https://github.com/apollographql/federation-rs/pull/236 in
`router-bridge@0.1.13-beta.0+v2.3.0-beta.3`.

However, an underlying dependency, `deno_core`, has pinned a lot of
dependencies in their latest release to versions which predate the
versions which we now have in `apollo-router`, which we have updated to
via `cargo update` and various Renovate-automated upgrades.

Therefore, to make `0.1.13-beta.0+v2.3.0-beta.3` work, , we need to
revert some of those dependency updates to allow the version
dependencies to be resolved.

Co-authored-by: Jesse Rosenberger <git@jro.cc>