Fix behaviour when reaching the maximum number of file descriptors (#2395)
- Fix #2384

When the number of connections increases, we might reach the maximum
number of file descriptors. At this point, trying to accept a new
connection results in an error, and the router tries in a loop to accept
connections.
What makes the behaviour worse:
- calling the `accept()` syscall makes the current thread block
- since we're not accepting connections, the accept queue stays full so
the syscall will never return the "would block" error
- if the router only has one core available, this leaves no time for the
rest of the router to work so it completely hangs (CTRL+C does not even
work)

This PR will fix the behaviour of the router when reaching this limit:
- [x] reduce the number of log messages that will appear
- [x] add a metric tracking the number of client sessions
- [ ] back off on accepting new connections until we have more file
descriptors available

That last point will require some thought. Right now there's a small
delay (1ms) to let the router perform other work, but it will go in a
loop of accepting new connections, sleeping because it reaches the max,
then accepting again, so some kind of hysteresis mechanism should be
used to get some breathing room.

Manually limiting the maximum number of client sessions (#2385) would
give us more flexibility:
- we can set it lower than the max open file limit, to leave enough file
descriptors for backend connections
- we can check at startup that it is indeed lower than max open file
(and prevent it from booting if not)
- we can use a design similar to
[BandAid](https://dropbox.tech/infrastructure/meet-bandaid-the-dropbox-service-proxy):
- accept connections as soon as possible, put them in a LIFO queue with
a configurable timeout. Older connections get closed
- if we reach the maximum number of connections, we know we cannot
handle the new traffic, so we aggressively close new connections as soon
as they are accepted. This signals early to the client that they should
retry with another server

Co-authored-by: Jesse Rosenberger <git@jro.cc>