Supports federation 2.3 in the router
The query planner in Federation 2.3 adds a new concept of "data
rewrites" for fetches in order to support both `@interfaceObject` and to
support the fix for https://github.com/apollographql/federation/issues/1257.

Those "rewrites" describe simple updates that need to be performed
either on the inputs (the "representations" passed to `_entities`; need
to rewrite the `__typename` when sending queries to an `@interfaceObject`)
or the output of a fetch (needed when a field has been aliased to permit
the subgraph query to be valid, but that field needs to be "un-aliased"
to its original name after the fetch).

This commit implements those rewrites.