chore(deps): update `libgit2-sys`, again.
This follows up https://github.com/apollographql/router/pull/2457 and
https://github.com/apollographql/router/pull/2497 which
both attempted to update `libgit2-sys`.

The changeset of https://github.com/apollographql/router/pull/2503
inadvertently resulted in a reversion of that version bump, which brought
back the offending version and upset our compliance check.

Thankfully, we have tooling that catches this stuff (even if it wasn't on
the PR itself).