chore: Phase 1 of adopting a new changelog pattern
Our `NEXT_CHANGELOG.md` hasn't scaled well, which is not entirely suprising
considering how many concurrently in-flight PRs and regular releases we
have.  In most cases, we fight merge conflicts daily.

Other project have [solved this
problem](https://github.com/changesets/changesets), but they're not written
in Rust and we're not sure if we need the full force of Node.js in our
projects, even if adopting it out of the box is tempting (at the cost of
otherwise no-op `package.json`s appearing throughout our repo).

For now, the motivation is to give us a place to start putting changelogs
that is NOT `NEXT_CHANGELOG.md` and the
multiple-files-in-one-directory-that-are-slurped-up-at-release-time approach
seems like a swell one that will avoid the biggest pain: merge conflicts.

The next step is naturally going to be `xtask` tooling, but I'll save that for a
follow-up PR.  For now, the `README.md` has the instructions on how to
create files in this directory.