chore: Phase 2 of changelog project, now w/ optionally prepopulated CHANGELOG (#2545)
This follows up on Phase 1 of the project I started in #2534 by bringing
along a tool that lets one stub out the nessary changeset file.

This stubbing takes place through the use of a new command:

    cargo xtask changeset create

There are no required arguments to pass, though it's possible to specify
some arguments that will shortcut interactive prompts. For example, the
use of the `--class <class>` (where `<class>` is one of `feat`, `maint`,
etc.) will skip the interactive prompt asking for precisely that.

In most cases, the prompt has sensible defaults. Additionally, the tool
will attempt to pre-populate the changeset entry with information by
matching the branch name the developer is currently working on against
any PRs from that branch on GitHub which have a pull request open
already.

This means that a very practical work-flow for using this command is to:

- Develop a change
- Push up a PR _without_ a changeset
- Run `cargo xtask changeset create` and follow the prompts
- Add and commit the resulting file in the `.changesets/` directory.
- Push that up to GitHub on the same branch

---------

Co-authored-by: Jeremy Lempereur <jeremy.lempereur@iomentum.com>