fix: `__typename` when formatting response in some `@interfaceObject` cases
The introduction of `@interfaceObject` in federation 2.3 has for
consequence that a subgraph (that uses an `@interfaceObject`) may return
in its response a `__typename` that corresponds to an interface type in
the supergraph (but which, locally to that subgraph, is an object type).

The idea is that if that `__typename` is requested, then the query
planner will ensure that another follow-up fetch will override that
value so it maps to a proper object type.

However, in some cases, that `__typename` is not queried, nor is there
another reason to resolve the actual underlying implementation type of
the underlying object (meaning that only fields of the interface are
queried), and in that case the `__typename` may still point to the
interface type when `format_response` is called. Unfortunately, that
method currently nullify the whole object in such case, which is not the
behaviour we want and was not caught in #2489.