Add support for using JWKS which don't specify "alg" (#2540)
- Fix #912

Many Identity Providers (IdP) don't specify "alg" in the keys of their
provided JWKSets. For example: Azure, Ping, etc...

This is fine as it isn't part of the specification. However, it provides
a challenge when it comes to JWT verification, because we need to know
which algorithm to use to verify the JWT.

This enhancement adds support for searching through the available
JWKSets in the router, trying to find the most suitable key for JWT
verification. Ideally, we will have a key which has both a matching
"kid" and an "alg" which we can use for verification. In the absence of
either or both of those, we may still be able to find a key which we can
use to verify our JWT as follows:
- If we find a key which has a matching "kid" and the correct family of
algorithm we will favour that first.
- If we find a key which has the correct family of algorithm, we will
favour that second.

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    ~- [ ] Integration Tests~
    ~- [ ] Manual Tests~

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`